requirement.txt
absl-py==1.2.0
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.2
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
bidict==0.22.1
billiard==3.6.4.0
bleach==5.0.1
blinker==1.4
boto3==1.26.69
botocore==1.29.69
Brlapi==0.7.0
cachetools==5.2.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
coloredlogs==15.0.1
colorlog==6.7.0
command-not-found==0.3
compiler==0.2.0
configparser==5.3.0
construct==2.5.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.30
dataclasses-json==0.5.7
dateparser==1.1.7
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.4
distro==1.7.0
distro-info===0.23ubuntu1
Django==4.1.5
django-admin-soft-dashboard==1.0.12
django-cors-headers==3.13.0
django-crispy-forms==1.8.1
django-crontab==0.7.1
django-debug-toolbar==3.8.1
django-directory-tree==1.7
django-el-pagination==3.3.0
django-elasticsearch-dsl==7.3
django-environ==0.9.0
django-livereload==1.7
django-multiselectfield==0.1.12
django-online-users==0.3
django-pandas==0.6.6
django-redis==5.2.0
django-rest-auth==0.9.5
django-session-timeout==0.1.0
django-smuggler==1.0.4
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
easygui==0.98.3
elasticsearch==7.17.9
elasticsearch-dsl==7.4.0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==0.10.0
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.7.1
Flask==2.2.2
Flask-Cors==3.0.10
flask-sock==0.6.0
Flask-SocketIO==5.3.4
flatbuffers==23.3.3
fonttools==4.33.3
fpdf==1.7.2
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.2
gast==0.4.0
GDAL==3.0.4
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.47.0
gunicorn==20.0.2
h11==0.9.0
h5py==3.8.0
haystack==0.42
html5tagger==1.3.0
httplib2==0.22.0
httptools==0.1.2
huggingface-hub==0.15.1
humanfriendly==10.0
idna==2.8
imbalanced-learn==0.8.1
imblearn==0.0
imgviz==1.5.1
importlib-metadata==4.12.0
importlib-resources==5.10.2
imutils==0.5.4
ipykernel==6.17.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
iso8601==1.0.2
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==7.4.1
jupyter-core==5.0.0
jupyter-events==0.6.0
jupyter-packaging==0.12.3
jupyter-server==1.23.4
jupyter-server-terminals==0.4.4
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.18.0
jupyterlab-widgets==3.0.7
keras==2.12.0
keyring==18.0.1
kiwisolver==1.4.3
kneed==0.8.1
kombu==5.2.4
labelme==5.0.5
langchain==0.0.160
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==16.0.0
livereload==2.6.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.2.2
matplotlib-inline==0.1.6
maxminddb==2.2.0
mistune==2.0.4
ml-dtypes==0.1.0
monotonic==1.5
more-itertools==4.2.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.31
mysqlclient==2.1.1
natsort==8.2.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
netifaces==0.10.4
nltk==3.8.1
notebook==6.5.2
notebook-shim==0.2.2
npm==0.1.1
numexpr==2.8.4
numpy==1.23.0
oauthlib==3.1.0
onnx==1.13.1
onnxruntime==1.14.1
open3d==0.13.0
openai==0.27.7
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
optional-django==0.1.0
outcome==1.2.0
packaging==21.3
pandas==1.4.4
pandas-stubs==2.0.2.230605
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
patsy==0.5.2
pefile==2023.2.7
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.4.0
pivottablejs==0.9.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
plotly==5.14.1
pmdarima==1.8.5
prometheus-client==0.15.0
prompt-toolkit==3.0.30
protobuf==4.22.3
psutil==5.9.1
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
pycups==1.9.73
pydantic==1.10.8
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==2.6.0
pymacaroons==0.13.0
pymongo==4.2.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.6
python-dotenv==0.21.0
python-engineio==4.4.1
python-json-logger==2.0.4
python-ptrace==0.9.8
python-socketio==5.8.0
python-vlc==3.0.18121
python3-openid==3.2.0
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==6.0
pyzmq==24.0.1
qimage2ndarray==1.10.0
QtPy==2.2.1
ratelim==0.1.6
raven==6.10.0
reactpy==1.0.0
redis==4.5.1
regex==2022.10.31
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
reverse-geocode==1.4.1
reverse-geocoder==1.5.1
rsa==4.9
s3transfer==0.6.0
sanic==23.3.0
Sanic-Cors==2.2.0
sanic-routing==22.8.0
scikit-learn==1.0.2
scipy==1.10.1
seaborn==0.11.1
SecretStorage==2.3.1
selenium==4.10.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.23.1
serial==0.0.97
simple-websocket==0.10.1
simplejson==3.16.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.9.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
speedtest==0.0.1
SQLAlchemy==2.0.15
sqlparse==0.4.2
stack-data==0.4.0
standard-scaler==0.3
statsmodels==0.13.5
style==1.1.0
suds==1.1.2
sympy==1.12
systemd-python==234
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.1.0
terminado==0.17.1
texttable==1.6.4
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.11.8
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
tracerite==1.1.0
traitlets==5.3.0
transformers==4.29.2
trio==0.22.0
trio-websocket==0.10.3
types-pytz==2023.3.0.0
typing-extensions==4.3.0
typing-inspect==0.9.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.7.0
ultralytics==8.0.105
unattended-upgrades==0.1
Unipath==1.1
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.12
usb-creator==0.3.7
uvicorn==0.11.3
uvloop==0.17.0
vine==5.0.0
virtualenv==20.0.17
voila==0.4.0
wadllib==1.3.3
wcwidth==0.2.5
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==0.59.0
websockets==10.4
Werkzeug==2.2.2
whitenoise==6.2.0
widgetsnbextension==4.0.7
winaudio==1.0.6
wrapt==1.14.1
wsproto==0.15.0
WTForms==2.1
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
